Patch from Terry Greeniaus:
We use the 64-bit tag field in the SRP commands to store information allowing us to track each command.  However, SCST only saves/uses 32-bits out of this 64-bit field.  This causes our code to fail when we get a response SRP command that doesn't match any tags we used.

I have updated SCST to work with our product by changing it to save all 64-bits of the tag.


git-svn-id: https://scst.svn.sourceforge.net/svnroot/scst/trunk@150 d57e44dd-8a1f-0410-8b47-8ef2f437770f