Synchronize with Feral CVS repository:
Make 2100/2200 and 2300 cards also maintain a target mode port database.

For the 24XX, we have to find the IID, N-port handle and channel when we
get an ATIO7. For the other cards, we already have the channel and N-port
handle and Port IID of the initiator, but have to find the S_ID.

Do the isp_find_pdb_by_{sid,loopid} functions search backwards to minimize
search time.

Don't load fw images more than once per device type!


git-svn-id: https://scst.svn.sourceforge.net/svnroot/scst/trunk@279 d57e44dd-8a1f-0410-8b47-8ef2f437770f