scst: Fix MAINTENANCE IN parsing
Although I do not know of any initiator that asks for the extended
parameter data format when submitting a REPORT TARGET PORT GROUPS
command, I think it's a good idea to support that data format (i.e.
three most significant bits of the second CDB byte are 001 instead
of 000). The attached patch should implement that. Additionally,
cmd->op_name is set to something more descriptive than
"MAINTENANCE IN" in that patch.

Signed-off-by: Bart Van Assche <bvanassche@acm.org>



git-svn-id: https://scst.svn.sourceforge.net/svnroot/scst/trunk@4385 d57e44dd-8a1f-0410-8b47-8ef2f437770f